Don't linkify profs in section table on search page
Fixes #102

Currently there are broken links on the search page ("Explore courses") due
to #98. This is the simplest fix -- just don't link up those profs. Ideally,
we make those links look like '/course/econ101#l_smith' so that a user could
click directly to the prof review on the course details page, but due to the
way we async-load prof reviews, this won't actually work.